requirements.txt
mysql-connector-python==9.1.0
Flask==3.0.3